requirements.txt
streamlit>=1.28.0
fastapi>=0.100.0
uvicorn>=0.20.0
groq>=0.4.0
pydantic>=2.0.0
python-dotenv>=1.0.0
requests>=2.28.0
sentence-transformers>=2.2.0
numpy>=1.21.0
pandas>=1.5.0
scikit-learn>=1.0.0
python-multipart>=0.0.5
httpx>=0.24.0
plotly>=5.0.0
matplotlib>=3.5.0
